requirements.txt
pandas==2.3.1
openpyxl==3.1.5
requests==2.32.4
boto3==1.39.17
python-dotenv==1.1.1
pyarrow==21.0.0